Move bbcvertex and associated map to reco package area
