+++ offline/packages/bbc/Makefile.am
@@ -27,7 +27,11 @@ pkginclude_HEADERS = \
   BbcPmtContainerV1.h \
   BbcPmtHit.h \
   BbcPmtHitV1.h \
-  BbcReturnCodes.h
+  BbcReturnCodes.h \
+  BbcVertex.h \
+  BbcVertexv1.h \
+  BbcVertexMap.h \
+  BbcVertexMapv1.h
 
 ROOTDICTS = \
   BbcOut_Dict.cc \
@@ -37,7 +41,11 @@ ROOTDICTS = \
   BbcPmtHit_Dict.cc \
   BbcPmtHitV1_Dict.cc \
   BbcPmtContainer_Dict.cc \
-  BbcPmtContainerV1_Dict.cc
+  BbcPmtContainerV1_Dict.cc \
+  BbcVertex_Dict.cc \
+  BbcVertexv1_Dict.cc \
+  BbcVertexMap_Dict.cc \
+  BbcVertexMapv1_Dict.cc
 
 pcmdir = $(libdir)
 
@@ -49,7 +57,11 @@ nobase_dist_pcm_DATA = \
   BbcPmtHit_Dict_rdict.pcm \
   BbcPmtHitV1_Dict_rdict.pcm \
   BbcPmtContainer_Dict_rdict.pcm \
-  BbcPmtContainerV1_Dict_rdict.pcm
+  BbcPmtContainerV1_Dict_rdict.pcm \
+  BbcVertex_Dict_rdict.pcm \
+  BbcVertexv1_Dict_rdict.pcm \
+  BbcVertexMap_Dict_rdict.pcm \
+  BbcVertexMapv1_Dict_rdict.pcm
 
 
 libbbc_io_la_SOURCES = \
@@ -61,7 +73,10 @@ libbbc_io_la_SOURCES = \
   BbcPmtHit.cc \
   BbcPmtHitV1.cc \
   BbcPmtContainer.cc \
-  BbcPmtContainerV1.cc
+  BbcPmtContainerV1.cc \
+  BbcVertexv1.cc \
+  BbcVertexMap.cc \
+  BbcVertexMapv1.cc
 
 libbbc_io_la_LIBADD = \
   -lphool
